Permisos de editor

+++ resources/views/usuarios/index.blade.php
@@ -5,10 +5,6 @@
 @endsection
 
 @section('content-center')
-    <div class="text-center">
-        <img class="col-4 img-thumbnail" src="{{ asset('storage/usuarios/' . $usuario->id . '.png') }}" onerror="this.style.display='none'">
-    </div>
-
     <h1 class="col-12 text-center mt-3">{{ $usuario->nombre }} {{ $usuario->apellido_paterno }} {{ $usuario->apellido_materno }}</h1>
 
     <label class="text-muted mt-4">Correo electrónico: {{ $usuario->email }}</label>
@@ -29,9 +25,11 @@
 
                     @slot('route', route('publicaciones.show', $pub))
 
-                    <a class="btn btn-secondary ms-2" href="{{ route('publicaciones.edit', $pub) }}">Editar</a>
+                    @if(Auth::user()->editor)
+                        <a class="btn btn-secondary ms-2" href="{{ route('publicaciones.edit', $pub) }}">Editar</a>
 
-                    <button class="btn btn-danger ms-2" type="button" onclick="eliminarPublicacion('{{ route('publicaciones.destroy', $pub) }}', '{{ $pub->titulo }}', {{ count($pub->respuestas) }});">Eliminar</button>
+                        <button class="btn btn-danger ms-2" type="button" onclick="eliminarPublicacion('{{ route('publicaciones.destroy', $pub) }}', '{{ $pub->titulo }}', {{ count($pub->respuestas) }});">Eliminar</button>
+                    @endif
                 @endcomponent
             @endforeach
         </div>
@@ -57,7 +55,9 @@
 @endsection
 
 @section('content-right')
-    <a class="btn btn-primary" href="{{ route('publicaciones.create') }}">Nueva publicación</a>
+    @if(Auth::user()->editor)
+        <a class="btn btn-primary" href="{{ route('publicaciones.create') }}">Nueva publicación</a>
+    @endif
 @endsection
 
 @section('content-other')
